requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
torch
transformers
scipy
numpy
redis
llama-cpp-python
tqdm
